Move Loc and ILoc classes into their own file indexing.py
