+++ python/cudf/dataframe/dataframe.py
@@ -41,6 +41,7 @@ from cudf.dataframe.buffer import Buffer
 from cudf.bindings.nvtx import nvtx_range_push, nvtx_range_pop
 from cudf._sort import get_sorted_inds
 from cudf.dataframe import columnops
+from cudf.indexing import Loc, Iloc
 
 import cudf.bindings.join as cpp_join
 import cudf.bindings.hash as cpp_hash
@@ -2797,157 +2798,6 @@ class DataFrame(object):
         """{docstring}"""
         import cudf.io.dlpack as dlpack
         return dlpack.to_dlpack(self)
-
-
-class Loc(object):
-    """
-    For selection by label.
-    """
-
-    def __init__(self, df):
-        self._df = df
-
-    def __getitem__(self, arg):
-
-        if type(arg) is not tuple:
-            arg = (arg, self._df.columns)
-
-        if self._is_multiindex_arg(arg):
-            return self._getitem_multiindex_arg(arg)
-        elif self._is_scalar_access(arg):
-            return self._getitem_scalar(arg)
-
-        df = self._getitem_tuple_arg(arg)
-
-        if self._can_downcast_to_series(df):
-            return self._downcast_to_series(df, arg)
-
-        return df
-
-    def _getitem_tuple_arg(self, arg):
-        columns = self._get_column_selection(arg[1])
-        df = DataFrame()
-        for col in columns:
-            df.add_column(name=col, data=self._df[col].loc[arg[0]])
-        return df
-
-    def _getitem_scalar(self, arg):
-        return self._df[arg[1]].loc[arg[0]]
-
-    def _getitem_multiindex_arg(self, arg):
-        # Explicitly ONLY support tuple indexes into MultiIndex.
-        # Pandas allows non tuple indices and warns "results may be
-        # undefined."
-        return self._df._index._get_row_major(self._df, arg)
-
-    def _normalize_dtypes(self, df):
-        dtypes = df.dtypes.values.tolist()
-        normalized_dtype = np.result_type(*dtypes)
-        for name, col in df.iteritems():
-            df[name] = col.astype(normalized_dtype)
-        return df
-
-    def _is_multiindex_arg(self, arg):
-        return (
-            isinstance(self._df.index, cudf.dataframe.multiindex.MultiIndex)
-            and isinstance(arg, tuple)
-        )
-
-    def _get_column_selection(self, arg):
-        if self._is_single_value(arg):
-            return [arg]
-        elif isinstance(arg, slice):
-            start, stop, step = arg.indices(self._df.shape[1])
-            cols = []
-            for i in range(start, stop, step):
-                cols.append(self._df.columns[i])
-            return cols
-        else:
-            return arg
-
-    def _is_single_value(self, val):
-        from pandas.core.dtypes.common import is_datetime_or_timedelta_dtype
-        if (
-                isinstance(val, str)
-                or isinstance(val, numbers.Number)
-                or is_datetime_or_timedelta_dtype(val)
-                or isinstance(val, pd.Timestamp)
-                or isinstance(val, pd.Categorical)
-        ):
-            return True
-        return False
-
-    def _is_scalar_access(self, arg):
-        if isinstance(arg, str):
-            return False
-        if not hasattr(arg, '__len__'):
-            return False
-        for obj in arg:
-            if not self._is_single_value(obj):
-                return False
-        return True
-
-    def _can_downcast_to_series(self, df):
-        nrows, ncols = df.shape
-        if nrows == 1:
-            dtypes = df.dtypes.values.tolist()
-            all_numeric = all([pd.api.types.is_numeric_dtype(t) for t in dtypes])
-            all_identical = dtypes.count(dtypes[0]) == len(dtypes)
-            if all_numeric or all_identical:
-                return True
-        if ncols == 1:
-            return True
-        return False
-
-    def _downcast_to_series(self, df, args):
-        if df.shape[0] == 1:
-            indices = df.columns
-            df = self._normalize_dtypes(df)
-            sr = df.T
-            sr._rename_columns([args[0]])
-            sr = sr.set_index(df.columns)
-            return sr[args[0]]
-
-        if df.shape[1] == 1:
-            return df[args[1]]
-
-
-class Iloc(object):
-    """
-    For integer-location based selection.
-    """
-
-    def __init__(self, df):
-        self._df = df
-
-    def __getitem__(self, arg):
-        if isinstance(arg, (tuple)):
-            if len(arg) == 1:
-                arg = list(arg)
-            elif len(arg) == 2:
-                return self[arg[0]][arg[1]]
-            else:
-                return pd.core.indexing.IndexingError(
-                    "Too many indexers"
-                )
-
-        if isinstance(arg, numbers.Integral):
-            rows = []
-            for col in self._df.columns:
-                rows.append(self._df[col][arg])
-            return Series(np.array(rows), name=arg)
-        else:
-            df = DataFrame()
-            for col in self._df.columns:
-                df[col] = self._df[col][arg]
-            df.index = self._df.index[arg]
-
-            return df
-
-    def __setitem__(self, key, value):
-        # throws an exception while updating
-        msg = "updating columns using iloc is not allowed"
-        raise ValueError(msg)
 
 
 def from_pandas(obj):
